v8.2.1: Complete no-reward patient releases
