+++ .github/fixtures/main-style-source-headroom.json
@@ -101,24 +101,24 @@
   },
   "v8Candidate": {
     "issue": 565,
-    "version": "8.2.0",
-    "sourceBytes": 1658510,
-    "sourceLines": 25335,
-    "sourceSha256": "865598afd3546b8a5604a223408f1618f59b11c460a0858ac9e2c45b906df91a",
+    "version": "8.2.1",
+    "sourceBytes": 1662753,
+    "sourceLines": 25418,
+    "sourceSha256": "71953aedd5408612bb4f2ae42f7896d01ee48ce2ab732c12d5617485d4bf0511",
     "templateBytes": 683558,
     "templateLines": 7858,
     "templateSha256": "8be4b32d5d4a001e15a5606ac28238fef22909e3276d791a3350851611602226",
     "canonicalCssSha256": "b1bf22273aa822f6966504771ee3a37210c32381d045ed374d7689992a1cd164",
-    "maxSourceBytes": 1678510,
-    "maxSourceLines": 25585,
+    "maxSourceBytes": 1682753,
+    "maxSourceLines": 25668,
     "minTemplateBytes": 500000,
-    "baseline": "8.1.5",
+    "baseline": "8.2.0",
     "approvedGrowth": {
-      "sourceBytes": 13966,
-      "sourceLines": 292,
+      "sourceBytes": 18209,
+      "sourceLines": 375,
       "templateBytes": 0,
       "templateLines": 0
     },
-    "scope": "Issue #565 verified-vehicle no-reward patient release path with same-mission reopening, per-patient verification, repeated-click protection and native discharge fallback"
+    "scope": "Issue #565 completion-aware same-origin no-reward release, deferred-control wait, same-vehicle patient-count verification and browser-faithful regression"
   }
 }
