--- .github/workflows/temporary-issue565-v8-2-1-executor.yml
@@ -1,105 +1,0 @@
-name: Temporary Issue 565 v8.2.1 Executor
-
-on:
-  pull_request:
-    types: [opened, reopened, synchronize]
-
-permissions:
-  contents: read
-  pull-requests: read
-
-concurrency:
-  group: temporary-issue565-v8-2-1
-  cancel-in-progress: false
-
-jobs:
-  execute:
-    if: >-
-      github.repository == 'Conroy1988/missionchief-toolkit-assets' &&
-      github.event.pull_request.number == 570 &&
-      github.event.pull_request.head.ref == 'fix/transport-sweep-release-navigation-v8-2-1'
-    runs-on: ubuntu-latest
-    timeout-minutes: 40
-    steps:
-      - name: Require owner branch token
-        env:
-          OWNER_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          [[ -n "$OWNER_TOKEN" ]] || { echo 'DEVELOPMENT_PR_TOKEN is required.' >&2; exit 1; }
-
-      - name: Check out exact pull-request head
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: ${{ github.event.pull_request.head.sha }}
-          fetch-depth: 0
-          token: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-
-      - name: Apply, validate, clean and publish
-        env:
-          EXPECTED_MAIN: 1655ef4316e7e677857dcbf7e5b170671fe6c73b
-          EXPECTED_HEAD: ${{ github.event.pull_request.head.sha }}
-          BRANCH: fix/transport-sweep-release-navigation-v8-2-1
-          PACKAGE: .github/development-packages/issue565_transport_sweep_release_v8_2_1.py
-          NORMALIZER: .github/development-packages/issue565_v821_source_indent_fix.py
-          NATIVE_FIX: .github/development-packages/issue565_v821_native_contract_fix.py
-          PERFORMANCE_FIX: .github/development-packages/issue565_v821_performance_history_fix.py
-          EXECUTOR: .github/workflows/temporary-issue565-v8-2-1-executor.yml
-        shell: bash
-        run: |
-          set -euo pipefail
-          exec > >(tee "$RUNNER_TEMP/issue565-v821.log") 2>&1
-          echo stage=verify-carrier
-          git fetch --no-tags origin main
-          test "$(git rev-parse origin/main)" = "$EXPECTED_MAIN"
-          test -f "$PACKAGE"
-          test -f "$NORMALIZER"
-          test -f "$NATIVE_FIX"
-          test -f "$PERFORMANCE_FIX"
-          test -f "$EXECUTOR"
-
-          echo stage=apply-package
-          python3 "$PACKAGE"
-          python3 "$NORMALIZER"
-          python3 "$NATIVE_FIX"
-          python3 "$PERFORMANCE_FIX"
-
-          echo stage=install-browser-runtime
-          npm install --no-save --package-lock=false --ignore-scripts jsdom@26.1.0
-
-          echo stage=targeted-regression
-          python3 .github/scripts/test_issue565_transport_sweep_no_reward.py
-          node .github/scripts/test_issue565_transport_sweep_no_reward_runtime.mjs
-
-          echo stage=canonical-validation
-          python3 .github/scripts/validate_userscript.py
-          node --check src/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent src/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent src/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-
-          echo stage=full-preflight
-          bash .github/scripts/run_userscript_preflight.sh --all
-
-          echo stage=clean-carrier
-          rm -rf node_modules
-          rm -f "$PACKAGE" "$NORMALIZER" "$NATIVE_FIX" "$PERFORMANCE_FIX" "$EXECUTOR"
-          find .github/development-packages -type d -empty -delete 2>/dev/null || true
-          test -z "$(git ls-files node_modules)"
-
-          echo stage=publish-candidate
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo 'No candidate changes produced.' >&2; exit 1; }
-          git commit -m "v8.2.1: Complete no-reward patient releases"
-          git push --force-with-lease="refs/heads/${BRANCH}:${EXPECTED_HEAD}" origin "HEAD:${BRANCH}"
-
-      - name: Upload executor diagnostics
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue565-v8-2-1-executor-${{ github.sha }}
-          path: ${{ runner.temp }}/issue565-v821.log
-          if-no-files-found: warn
-          retention-days: 30
