+++ docs/issue-565-transport-sweep-no-reward.md
@@ -1,7 +1,7 @@
 # Issue #565 — Patient Transport Sweep no-reward release path
 
-Toolkit v8.2.0 recognises only the exact visible `Release patient (No reward)` mission control whose same-origin path matches `/vehicles/{vehicleId}/patient/-1`.
+Toolkit v8.2.1 recognises only the exact visible `Release patient (No reward)` control whose same-origin path matches `/vehicles/{vehicleId}/patient/-1` and whose vehicle is already verified by the existing sweep candidate collector.
 
-Only vehicle IDs already verified by the existing sweep candidate collector are eligible. The sweep releases one patient, reopens the same mission, verifies that the released vehicle-specific control is absent, records one confirmed patient outcome and then repeats for the next patient. A persistent control, failed mission reopen or cancellation stops the fast path safely. When no matching control exists, the established MissionChief-native vehicle-window discharge process remains unchanged.
+The control may be inserted after mission render, so the sweep waits boundedly before selecting native fallback. It then completes one same-origin GET using the exact inspected href, waits for the response body, reopens the mission and verifies the patient count in that vehicle row has decreased. A vehicle carrying several patients may therefore be released repeatedly (`3 → 2 → 1 → 0`) with a unique confirmation identity for every patient.
 
-The implementation adds no observer, interval or new network-request call site. Permanent executable coverage includes multiple patients, allowance limiting, missing controls, persistent controls, reopen failure, cancellation and mission-progress isolation.
+A failed request, unchanged patient count, cancellation or failed mission reopen stops the optional path safely. The established MissionChief-native vehicle-window discharge process remains the fallback. The correction adds one user-invoked network-request site and no observer, interval or disabled-state work.
